pyproject.toml
[project]
name = "discord-bot"
version = "0.1.0"
description = "Discord bot with tier list, economy, and moderation features"
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.3.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.8.0",
    "pillow>=10.0.0",
    "psutil>=7.0.0",
    "aiosqlite>=0.19.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
